Add a function to handle storing multiple regions
It will be used to track the which memory is accessible by a capability
in CHERI. It allows for new regions to be added to existing regions and
have them merge or split as needed. The permissions are stored as the
union of each added region, so if one was read only, and another was just
for writing it will be recorded as read-write.

It ignores the max_perms fiels as this will be removed soon. There will
also be future work to add a callback for when new regions are added and
when an existing region is changed, i.e. possible new memory to scan.